// src/api/openAPI/models/continent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SMAIAX Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const Continent = {
    Africa: 'Africa',
    Antarctica: 'Antarctica',
    Asia: 'Asia',
    Europe: 'Europe',
    NorthAmerica: 'NorthAmerica',
    Oceania: 'Oceania',
    SouthAmerica: 'SouthAmerica'
} as const;

export type Continent = typeof Continent[keyof typeof Continent];